moved encoder stuff to callbacks.cpp
